sync

+++ BookParser/Pages/MainWindowModel.cs
@@ -1,6 +1,4 @@
-﻿using GalaSoft.MvvmLight;
-using GalaSoft.MvvmLight.Command;
-using Microsoft.WindowsAPICodePack.Dialogs;
+﻿using BookParser.Folders;
 using ParserEngine.Engine;
 using ParserEngine.Models;
 using PropertyChanged;
@@ -9,19 +7,37 @@ using System.Collections.Generic;
 using System.Collections.ObjectModel;
 using System.IO;
 using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+using System.Windows.Forms;
 using System.Windows.Input;
 
-namespace BookParser
+namespace BookParser.Pages
 {
     [AddINotifyPropertyChangedInterface]
-    public class MainWindowModel:ViewModelBase
+    public class MainWindowModel
     {
         private Website _website;
         private readonly CoreEngine CoreEngine;
         private readonly FileEngine FileEngine;
         public ObservableCollection<string> Logs { get; private set; }
 
-        public string Url { get; set; }
+        string url = string.Empty;
+        public string Url
+        {
+            get => url;
+            set
+            {
+                if (url != value)
+                {
+                    url = value;
+                    UpdateData(value);
+                }
+
+            }
+        }
+
+        public string StartPage { get; set; }
         public string UrlPlaceholder { get; set; }
         public string Author { get; set; }
         public string Bookname { get; set; }
@@ -31,7 +47,8 @@ namespace BookParser
         public ParserType HeaderParserType { get; set; }
         public string NextPageId { get; set; }
         public ParserType NextPageIdType { get; set; }
-        public string FolderPath { get; set; }
+        public ParserType ArbitaryType { get; set; }
+        public string ArbitaryInfo { get; set; }
         public Website Website
         {
             get => _website;
@@ -61,28 +78,48 @@ namespace BookParser
                 HeaderParserType = website.HeaderType;
                 NextPageId = website.NextData;
                 NextPageIdType = website.NextType;
+                ArbitaryInfo = website.ArbitaryData;
+                ArbitaryType = website.ArbitaryType;
             }
         }
         public ICommand ParseCommand => new RelayCommand(ParseAction);
-        public ICommand FolderCommand => new RelayCommand(FolderAction);
 
         public bool IsBusy { get; private set; }
 
-        private void FolderAction()
+
+        private async void UpdateData(string value)
         {
-            using (var dialog = new CommonOpenFileDialog())
+            var book = await CoreEngine.GetBookInfo(value);
+            if (book != null)
             {
-                dialog.IsFolderPicker = true;
-                if (dialog.ShowDialog() == CommonFileDialogResult.Ok)
-                {
-                    FolderPath = dialog.FileName;
-                }
+                Author = book.Author;
+                Bookname = book.Bookname;
+                StartPage = book.Url;
             }
+        }
+
+
+        private string FolderAction()
+        {
+            using var dialog = new FolderBrowserDialog
+            {
+                Description = "Time to select a folder",
+                UseDescriptionForTitle = true,
+                SelectedPath = Environment.GetFolderPath(Environment.SpecialFolder.DesktopDirectory)
+                              + Path.DirectorySeparatorChar,
+                ShowNewFolderButton = true
+            };
+
+            if (dialog.ShowDialog() == DialogResult.OK)
+            {
+                return dialog.SelectedPath;
+            }
+            else return string.Empty;
         }
 
         private async void ParseAction()
         {
-            if (string.IsNullOrEmpty(Url)) ShowMessage("Error", "Invalid Url");
+            if (string.IsNullOrEmpty(StartPage)) ShowMessage("Error", "Invalid Url");
             else if (string.IsNullOrEmpty(Author)) ShowMessage("Error", "Invalid Author");
             else if (string.IsNullOrEmpty(Bookname)) ShowMessage("Error", "Invalid bookname");
             else if (string.IsNullOrEmpty(ContentId))
@@ -91,10 +128,11 @@ namespace BookParser
                 ShowMessage("Error", "You must define the class/id of the next page");
             else if (string.IsNullOrEmpty(HeaderId))
                 ShowMessage("Error", "You must define the class/id of the next page");
-            else if (string.IsNullOrEmpty(FolderPath))
-                ShowMessage("Error", "You must define the export folder");
+
             else
             {
+                var folder = FolderAction();
+                if (string.IsNullOrWhiteSpace(folder)) return;
                 var path = Path.GetTempPath() + Bookname;
                 if (Directory.Exists(path)) Directory.Delete(path, true);
                 Directory.CreateDirectory(path);
@@ -106,7 +144,8 @@ namespace BookParser
                     FilePath = path,
                     NextChapterInfo = new ParseInfo(NextPageIdType, NextPageId),
                     TitleInfo = new ParseInfo(HeaderParserType, HeaderId),
-                    Url = Url
+                    ArbitaryInfo = new ParseInfo(ArbitaryType, ArbitaryInfo),
+                    Url = StartPage
                 };
                 Logs.Clear();
                 IsBusy = true;
@@ -114,7 +153,7 @@ namespace BookParser
                 var parseRes = await CoreEngine.Parse(book);
                 if (parseRes)
                 {
-                    var exportFile = Path.Combine(FolderPath, book.Bookname + ".epub");
+                    var exportFile = Path.Combine(folder, book.Bookname + ".epub");
                     await FileEngine.CreateBook(book, exportFile);
                 }
                 IsBusy = false;
